made changes

+++ js/main.js
@@ -7,7 +7,10 @@ function addTodo() {
         listItem.classList.add('all', 'active', 'todoItem');
         listItem.innerHTML = `
             <input type="checkbox" onchange="toggleCompleted(this)">
+            <div class="random">
             <span>${todoText}</span>
+            <i class="ri-delete-bin-7-line" onclick="removeTodo(this)"></i>
+            </div>
         `;
         // <button onclick="removeTodo(this)">Delete</button>
         const todoList = document.getElementById('todoList');
@@ -28,8 +31,8 @@ function toggleCompleted(checkbox) {
     }
 }
 
-function removeTodo(button) {
-    const listItem = button.parentNode;
+function removeTodo(i) {
+    const listItem = i.parentNode.parentNode;
     const todoList = document.getElementById('todoList');
     todoList.removeChild(listItem);
 }
@@ -45,6 +48,10 @@ function showAll() {
     const deleteAllbtn = document.getElementById('deleteAll');
     if(todoContainer.contains(deleteAllbtn))
         todoContainer.removeChild(deleteAllbtn);
+
+    const addTodoContainer = document.getElementById('addTodo');
+    if(addTodoContainer.style.display === 'none')
+        addTodoContainer.style.display = 'inline';
 }
 
 function showActive() {
@@ -60,19 +67,28 @@ function showActive() {
     const deleteAllbtn = document.getElementById('deleteAll');
     if(todoContainer.contains(deleteAllbtn))
         todoContainer.removeChild(deleteAllbtn);
+    const addTodoContainer = document.getElementById('addTodo');
+    if(addTodoContainer.style.display === 'none')
+        addTodoContainer.style.display = 'inline';
 }
 
 function showCompleted() {
-    const newButton = document.createElement('button');
-    newButton.textContent = 'Delete All';
-    newButton.id = 'deleteAll';
-    newButton.classList.add('deleteAll');
-    newButton.addEventListener('click', function() {
-        deleteAll();
-    });
-    const buttonContainer = document.getElementById('todoContainer');
-    buttonContainer.appendChild(newButton);
-
+    const todoContainer = document.getElementById('todoContainer');
+    const deleteAllbtn = document.getElementById('deleteAll');
+    if(!todoContainer.contains(deleteAllbtn))
+    {
+        const newButton = document.createElement('button');
+        newButton.id = 'deleteAll';
+        newButton.classList.add('deleteAll');
+        newButton.innerHTML = `
+            <i class="ri-delete-bin-7-line"></i>
+            delete all
+        `;
+        newButton.addEventListener('click', function() {
+            deleteAll(this);
+        });
+        todoContainer.appendChild(newButton);
+    }
     const ulElement = document.getElementById('todoList');
     const listItems = ulElement.getElementsByTagName('li');
     for (let i = 0; i < listItems.length; i++) {
@@ -82,9 +98,17 @@ function showCompleted() {
             listItems[i].classList.remove('hidden')
     }
     
+    const addTodoContainer = document.getElementById('addTodo');
+    if(!(addTodoContainer.style.display === 'none'))
+        addTodoContainer.style.display = 'none';
 
 }
 
 function deleteAll() {
-
+    const ulElement = document.getElementById('todoList');
+    const listItems = ulElement.getElementsByTagName('li');
+    for (let i = 0; i < listItems.length; i++) {
+        if(listItems[i].classList.contains('completed'))
+            todoList.removeChild(listItems[i]);
+    }
 }
